+++ api/migrations.py
@@ -102,6 +102,10 @@ class ModelMigration(DbMigration):  # pragma: no cover
                     doc['test_import_handler'] = None
                     self.collection.save(doc)
 
+    def allmigration10__add_current_task_id(self):
+        self.target = {'current_task_id': {'$exists': False}}
+        self.update = {'$set': {'current_task_id': ''}}
+
 
 class TestMigration(DbMigration):  # pragma: no cover
     DOC_CLASS = models.Test
@@ -121,11 +125,15 @@ class TestMigration(DbMigration):  # pragma: no cover
         self.target = {'created_by': {'$exists': False}}
         self.update = {'$set': {'created_by': {}}}
 
-    def allmigration04__add_examples_placement(self):
+    def allmigration04__add_current_task_id(self):
+        self.target = {'current_task_id': {'$exists': False}}
+        self.update = {'$set': {'current_task_id': ''}}
+
+    def allmigration05__add_examples_placement(self):
         self.target = {'examples_placement': {'$exists': False}}
         self.update = {'$set': {'examples_placement': 'Amazon S3'}}
 
-    def allmigration05__add_examples_fields(self):
+    def allmigration06__add_examples_fields(self):
         self.target = {'examples_fields': {'$exists': False}}
         self.update = {'$set': {'examples_fields': []}}
 
@@ -177,6 +185,10 @@ class DataSetMigration(DbMigration):  # pragma: no cover
     def allmigration06__add_on_s3(self):
         self.target = {'on_s3': {'$exists': False}}
         self.update = {'$set': {'on_s3': False}}
+
+    def allmigration07__add_current_task_id(self):
+        self.target = {'current_task_id': {'$exists': False}}
+        self.update = {'$set': {'current_task_id': ''}}
 
 
 class InstanceMigration(DbMigration):  # pragma: no cover