+++ api/tasks.py
@@ -3,13 +3,14 @@ import json
 import logging
 import csv
 import uuid
-from itertools import izip, chain
+from itertools import izip
 from bson.objectid import ObjectId
 from os.path import exists
 from os import makedirs, system
 from datetime import timedelta, datetime
 from boto.exception import EC2ResponseError
 from celery.canvas import group
+from celery.signals import task_prerun, task_postrun
 
 from api import celery, app
 from api.models import Test, Model
@@ -538,9 +539,18 @@ def _add_example_to_mongo(test, vectorized_data, data, label, pred, prob):
     """
     model = test.model
     example = app.db.TestExample()
-    example.id = str(data.get(model.example_id, '-1'))
-    example.name = str(data.get(model.example_label, 'noname'))
-    example.label = str(label)
+    example_id = new_row.get(model.example_id, '-1')
+    try:
+        example['id'] = str(example_id)
+    except UnicodeEncodeError:
+        example['id'] = example_id.encode('utf-8')
+
+    example_name = new_row.get(model.example_label, 'noname')
+    try:
+        example['name'] = str(example_name)
+    except UnicodeEncodeError:
+        example['name'] = example_name.encode('utf-8')
+
     example.pred_label = str(pred)
     example.prob = prob.tolist()
     example.test = test
@@ -595,7 +605,6 @@ def store_examples(test_id, params):
             row = fp.readline()
             row = json.loads(row)
 
-            row = decode(row)
 
             example = app.db.TestExample.find_one({'_id': ObjectId(example_id)})
             if not example:
@@ -730,12 +739,62 @@ def get_csv_results(model_id, test_id, fields):
 def decode(row):
     from decimal import Decimal
     for key, val in row.iteritems():
-        try:
-            if isinstance(val, basestring):
-                row[key] = val.encode('ascii', 'ignore')
-            if isinstance(val, Decimal):
+        if isinstance(val, Decimal):
                 row[key] = val.to_eng_string()
-        except UnicodeDecodeError:
-            #logging.error('Error while decoding %s: %s', val, exc)
-            row[key] = ""
+        
+        if isinstance(val, basestring):
+            try:
+                #row[key] = val.encode('ascii', 'ignore')
+                row[key] = str(val)
+            # except UnicodeDecodeError:
+            #     #logging.error('Error while decoding %s: %s', val, exc)
+            #     row[key] = ""
+            except UnicodeEncodeError:
+                row[key] =val.encode('utf-8')
     return row
+
+
+@task_prerun.connect
+def task_prerun_handler(sender=None, task_id=None, task=None, args=None,
+                        kwargs=None, **kwds):  # pragma: no cover
+    cls = None
+    obj_id = None
+    if task.name == 'api.tasks.import_data':
+        cls = app.db.DataSet
+        obj_id = args[0] if len(args) else kwargs['dataset_id']
+    elif task.name == 'api.tasks.run_test':
+        cls = app.db.Test
+        obj_id = args[1] if len(args) else kwargs['test_id']
+    elif task.name in ('api.tasks.train_model',
+                       'api.tasks.fill_model_parameter_weights'):
+        cls = app.db.Model
+        if task.name == 'api.tasks.train_model':
+            obj_id = args[1] if len(args) else kwargs['model_id']
+        elif task.name == 'api.tasks.fill_model_parameter_weights':
+            obj_id = args[0] if len(args) else kwargs['model_id']
+
+    if cls and obj_id:
+        obj = cls.get_from_id(ObjectId(obj_id))
+        if obj:
+            obj.current_task_id = task_id
+            obj.save()
+
+
+@task_postrun.connect
+def task_postrun_handler(sender=None, task_id=None, task=None, args=None,
+                         kwargs=None, retval=None, state=None,
+                         **kwds):  # pragma: no cover
+    cls = None
+    if task.name == 'api.tasks.import_data':
+        cls = app.db.DataSet
+    elif task.name == 'api.tasks.run_test':
+        cls = app.db.Test
+    elif task.name in ('api.tasks.train_model',
+                       'api.tasks.fill_model_parameter_weights'):
+        cls = app.db.Model
+
+    if cls:
+        cls.collection.update(
+            {'current_task_id': task_id}, {'$set': {'current_task_id': ''}},
+            multi=True
+        )