+++ api/test/test_tasks.py
@@ -1,3 +1,5 @@
+# -*- coding: utf8 -*-
+
 from api.tasks import InvalidOperationError
 import os
 from bson import ObjectId
@@ -124,27 +126,25 @@ class TestTasksTests(BaseTestCase):
         import scipy
         from api.tasks import run_test
 
+        _fake_raw_data = [{'application_id': '123',
+                           'hire_outcome': '0',
+                           'title': 'A1'}] * 100
+
         class MetricsMock(MagicMock):
-            accuracy = 0.80
+            accuracy = 0.85
             classes_set = ['0', '1']
             _labels = ['0', '1'] * 50
 
-            METRICS_DICT = {
-                'confusion_matrix': [0, 0],
-                'roc_curve': [[0], [0], [0], [0]],
-                'precision_recall_curve': [[0], [0], [0], [0]],
-            }
-
             def get_metrics_dict(self):
-                return self.METRICS_DICT
+                return {
+                    'confusion_matrix': [0, 0],
+                    'roc_curve': [[0], [0], [0], [0]],
+                    'precision_recall_curve': [[0], [0], [0], [0]],
+                }
 
         def _fake_test(self, *args, **kwargs):
             _fake_test.called = True
-
-            self._raw_data = [{'application_id': '123',
-                               'hire_outcome': '0',
-                               'title': 'A1'}] * 100
-
+            self._raw_data = _fake_raw_data
             metrics_mock = MetricsMock()
             preds = Mock()
             preds.size = 100
@@ -236,6 +236,21 @@ class TestTasksTests(BaseTestCase):
             self.assertRaises(InvalidOperationError, run_test,
                               [self.dataset._id, ], self.test._id)
 
+
+            # Unicode encoding test
+            model.status = model.STATUS_TRAINED
+            model.save()
+            unicode_string = u'Привет!'
+            for row in _fake_raw_data:
+                row['opening_id'] = row['opening_title'] = unicode_string
+            self.db.TestExample.collection.remove(
+                {'test_id': str(self.test2._id)})
+            result = run_test([self.dataset._id, ], self.test2._id)
+            self.assertEquals(result, 'Test completed')
+            example = self.db.TestExample.find_one(
+                {'test_id': str(self.test2._id)})
+            self.assertEquals(example.id, unicode_string)
+            self.assertEquals(example.name, unicode_string)
 
     @mock_s3
     def test_store_examples(self):