+++ changelog.rst
@@ -3,9 +3,14 @@
 Changelog
 =========
 
+Relese 2013-09-??
+-----------------
+* Bug - Fixed deleting test and dataset while they is in progress
+
 
 Relese 2013-09-09
 -----------------
+* Bug - UnicodeEncodeError during run_test
 * Bug - Tests list should be updated after deleting test
 * Bug - Fixed dataset name editing
 * Bug - Remove links to dataset when deleting
@@ -16,6 +21,7 @@ Relese 2013-09-09
 * Task - Tests improvements
 * Taks - Adding dashboard page and API for getting some stat data
 * Taks - Humanize time format
+* Feature - Sort examples by probability
 * Feature - Adding filtering to models list
 * Feature - Added using exsiting file datset instead download from s3
 * Feature - Generating download from s3 dataset url only when user clicks download btn